업체 리스트 탭 완료

+++ PleaseSunshine/Models/LookCostData.swift
@@ -15,5 +15,5 @@ struct LookCostData: Codable {
 
 struct LookCost: Codable {
     let watt, savedMoney: Int
-    let installCostAvg, bePoint: Double
+    let installCostAvg, bePoint: Int
 }
